Alter glossary labels to pick up new theme styling when linked to
Version 0.14 of crate-docs-theme includes a special CSS style for links that
reference glossary entries. It does this by checking the `href` attribute for
the `#gloss-` string. For this to work, all labels used for terms in the
glossary should be prefixed with `gloss-`.

The `gloss-` prefix was chosen for the CrateDB Reference glossary because we
make heavy use of cross-referencing (i.e., the main docs repeatedly hyperlink
phrases to the relevant glossary term) and using `gloss-` instead of
`glossary-` cuts down on hard wrapping requirements (and also makes typing them
all a little easier).

I grepped the Cloud Reference docs for cross-references to the glossary, but I
could not find any outside of the glossary itself and the top-level
`index.rst`. Do you link to any of the defined terms from elsewhere? If so,
those links will need to be updated.

As an aside, we have found that cross-referencing the main docs to link to the
glossary is providing a lot of value through increased navigability and
discoverability. Check it out and see what you think:
<https://crate.io/docs/crate/reference/en/master/appendices/glossary.html>

With this change in place, links to defined terms will be styled black with a
dotted underline. This styling has been chosen to reduce visual noise and
indicate to the reader that a link goes to the glossary as opposed to a
location in the main documentation.

You may need to run `make reset` before you can pick up the theme changes.

This change also fixes a few minor RST style issues.